requirements.txt
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
colorama==0.4.6
colorgram.py==1.2.0
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prettytable==3.7.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.16.1
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.1
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
